Extract SurveyLines logic into its own class
